split up Expr.java
